+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - `gatekeeper_server`:
   - `login`: also respond with the server version.
 
+- base_codecs: ^1.0.1
+
+- pubspec_parse: ^1.5.0
+- path: ^1.9.1
+
 ## 1.0.3
 
 - New `AESEncryptor` and `ChainAESEncryptor`.